requirements.txt
gradio
python-docx
python-dotenv
langchain
langchain-core
langchain-community
langchain-google-genai
faiss-cpu
pypdf
docx2txt